+++ CHANGELOG.md
@@ -3,7 +3,13 @@ This project versioning adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 ### Added
+<<<<<<< HEAD
 - `WebDriverCheckboxes` and `WebDriverRadios` helper classes to simplify interaction with checkboxes and radio buttons.
+=======
+- Support for local storage and session storage via the new `WebDriver::storage()` method.
+
+---
+>>>>>>> duncan3dc/implement-web-storage
 
 ## 1.6.0 - 2018-05-16
 ### Added